fix f32 -> f64

+++ src/osu/pp.rs
@@ -578,7 +578,7 @@ impl OsuPpInner {
     
         let len_bonus = 1.1
             + 0.7 * (total_hits / 1500.0).min(1.0)
-            + (total_hits > 1500.0) as u8 as f32 * 0.7 * (total_hits / 1500.0).log10();
+            + (total_hits > 1500.0) as u8 as f64 * 0.7 * (total_hits / 1500.0).log10();
     
         aim_value *= len_bonus;
     
@@ -637,7 +637,7 @@ impl OsuPpInner {
     
         let len_bonus = 1.1
             + 0.7 * (total_hits / 1500.0).min(1.0)
-            + (total_hits > 1500.0) as u8 as f32 * 0.7 * (total_hits / 1500.0).log10();
+            + (total_hits > 1500.0) as u8 as f64 * 0.7 * (total_hits / 1500.0).log10();
         speed_value *= len_bonus;
     
         if self.effective_miss_count > 0.0 {
